Fix/Improve trimming system
Sprite.Width & Height are dimensions of initial image;
Sprite.SrcWidth & SrcHeight are dimensions of texture rectangle;

When drawing on screen or checking the size of a sprite for layout
purposes, use Width and Height as before;

When working with texture, always use ScrWidth and SrcHeight.

No-trim flag and implementation removed.